requirements.txt
asgiref==3.5.2
Django==3.2.13
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
mysqlclient==2.1.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
